fixed : 스크롤, 화면 초기위치조정
fixed : 스크롤, 화면 초기위치조정
