+++ client/src/pages/main/HomePage/styles/HomePageStyles.js
@@ -1,0 +1,8 @@
+import {
+  Borderdiv,
+  MainContainer,
+  MarginBottomTom,
+} from "./MainFrame/MainContainer";
+import { CategoryContainer } from "./CategoryContainerStyle/CategoryContainer";
+
+export { Borderdiv, MainContainer, MarginBottomTom, CategoryContainer };
