+++ client/src/pages/main/ProductsPage/styles/ProductPageStyle.jsx
@@ -1,0 +1,3 @@
+import { MainContainer, Borderdiv } from "./MainFrame/MainFrame";
+
+export { MainContainer, Borderdiv };
